Streamline dyna error handler
This preserves the behaviour that erroneous dynas break Vanilla's output
chain (in other words, the debug switch is needed to see the error
message). Even though the old cold looks like it tries to catch the
error and display a more friendly error message, it is broken in it's
use of disarm (which does not mutate).

Future changes may modify this behaviour and display an error message
in-line with the rest of the output. An even better idea would be to
support something like Firebug's debug console.